feat: interface repo layer
