+++ src/main/java/com/github/food2gether/sessionservice/repository/OrderRepository.java
@@ -2,9 +2,7 @@ package com.github.food2gether.sessionservice.repository;
 
 import com.github.food2gether.model.Order;
 import io.quarkus.hibernate.orm.panache.PanacheRepository;
-import jakarta.enterprise.context.ApplicationScoped;
 
-@ApplicationScoped
-public class OrderRepository implements PanacheRepository<Order> {
+public interface OrderRepository extends PanacheRepository<Order> {
 
 }
