+++ src/main/java/com/github/food2gether/sessionservice/repository/OrderRepositoryImpl.java
@@ -1,0 +1,8 @@
+package com.github.food2gether.sessionservice.repository;
+
+import jakarta.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
+public class OrderRepositoryImpl implements OrderRepository{
+
+}
